rm-username-form-view

+++ aircall_connector/models/call_history.py
@@ -37,7 +37,7 @@ class AirCall(models.Model):
     air_call_tag = fields.Many2many(comodel_name="res.partner.category", relation="call_tags_relation",
                                     column1="call_tag_id", column2="call_id", string="Tags", )
     notes = fields.Text(strng="Notes", required=False)
-    call_duration = fields.Integer(strng="Duration", required=False)
+    call_duration = fields.Float(strng="Duration", required=False)
     company_id = fields.Many2one('res.company', string='Company', default=lambda self: self.env.company)
 
     def write(self, values):
